Move arislog and vc-using-framestream to use vcpkg protobuf. (#51)
* Move arislogand vc-using-framestream to use vcpkg protobuf.
* Add SDK section on 'Verifying the Command Length Prefix'
* Add README note about use of vcpkg.
* Remove protobuf nuget from C++ projects.
* Reformat availability messages. Fixes #53.
* Update notes on vcpkg

Because vcpkg is a submodule and the protobuf package version is hard-coded within vcpkg we will not see vcpkg change unless **we** change it.

+++ sample-code/vc-using-framestream/vc-using-framestream/UdpListener.cpp
@@ -12,7 +12,7 @@ UdpListener::UdpListener(boost::asio::io_service & io, uint16_t port, bool reuse
   socket_.open(udp::v4());
   socket_.set_option(boost::asio::socket_base::reuse_address(reuseAddress));
   socket_.bind(udp::endpoint(udp::v4(), port));
-  socket_.set_option(boost::asio::socket_base::receive_buffer_size(recvBuffer_.capacity()));
+  socket_.set_option(boost::asio::socket_base::receive_buffer_size(static_cast<int>(recvBuffer_.capacity())));
 
   StartAsyncReceive();
 }
